Sandy bridge max flops per cycle was wrong
Updated the max flops per cycle. Which means all peak flops in pairwise_3d, pairwise_3d_histogram need to be multiplied by a factor of two. NOTE that, pairwise_3d_histogram will now show > 100% efficiency -- this is expected. I am brazenly saying that all pairs were evaluated for the 3d histogram but that is not the case. For a correct efficiency comparison, I have to calculate the actual number of pair distances computed. May be a job for the distant future...may be.